pull VRScene out of app into its own component
